ObjectId tests resurrected, comments added
